+++ SWP391_TEAM3_ShoesShop/nbproject/private/private.properties
@@ -1,7 +1,7 @@
-deploy.ant.properties.file=C:\\Users\\Tuan anh\\AppData\\Roaming\\NetBeans\\17\\tomcat101.properties
-j2ee.server.home=C:/Program Files/Apache Software Foundation/Tomcat 10.1_Tomcat
-j2ee.server.instance=tomcat101:home=C:\\Program Files\\Apache Software Foundation\\Tomcat 10.1_Tomcat
+deploy.ant.properties.file=C:\\Users\\Admin\\AppData\\Roaming\\NetBeans\\17\\tomcat101.properties
+j2ee.server.home=C:/Program Files/Apache Software Foundation/Tomcat 10.1
+j2ee.server.instance=tomcat101:home=C:\\Program Files\\Apache Software Foundation\\Tomcat 10.1
 javac.debug=true
 javadoc.preview=true
 selected.browser=default
-user.properties.file=C:\\Users\\Tuan anh\\AppData\\Roaming\\NetBeans\\17\\build.properties
+user.properties.file=C:\\Users\\Admin\\AppData\\Roaming\\NetBeans\\17\\build.properties
