--- SWP391_TEAM3_ShoesShop/src/java/Controllers/SearchProduct.java
@@ -1,114 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
- */
-package Controllers;
-
-import DAL.ProductDAO;
-import Models.Product;
-import java.io.IOException;
-import java.io.PrintWriter;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.util.List;
-
-/**
- *
- * @author thanh
- */
-@WebServlet 
-public class SearchProduct extends HttpServlet {
-
-    /**
-     * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
-     * methods.
-     *
-     * @param request servlet request
-     * @param response servlet response
-     * @throws ServletException if a servlet-specific error occurs
-     * @throws IOException if an I/O error occurs
-     */
-    protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        response.setContentType("text/html;charset=UTF-8");
-        try (PrintWriter out = response.getWriter()) {
-            /* TODO output your page here. You may use following sample code. */
-            out.println("<!DOCTYPE html>");
-            out.println("<html>");
-            out.println("<head>");
-            out.println("<title>Servlet SearchProduct</title>");            
-            out.println("</head>");
-            out.println("<body>");
-            out.println("<h1>Servlet SearchProduct at " + request.getContextPath() + "</h1>");
-            out.println("</body>");
-            out.println("</html>");
-        }
-    }
-
-    // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
-    /**
-     * Handles the HTTP <code>GET</code> method.
-     *
-     * @param request servlet request
-     * @param response servlet response
-     * @throws ServletException if a servlet-specific error occurs
-     * @throws IOException if an I/O error occurs
-     */
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        response.setContentType("text/html;charset=UTF-8");
-        request.setCharacterEncoding("UTF-8");
-        String txtSearch = request.getParameter("txt");
-        ProductDAO productDAO = new ProductDAO();
-        List<Product> list  = productDAO.searchByName(txtSearch);
-        PrintWriter out = response.getWriter();
-        for (Product o : list) {
-             out.println("<div class=\"product col-12 col-md-6 col-lg-4\">\n"
-                    + "                                <div class=\"card\">\n"
-                    + "                                    <img class=\"card-img-top\" src=\""+o.getAvatarP()+"\" alt=\"Card image cap\">\n"
-                    + "                                    <div class=\"card-body\">\n"
-                    + "                                        <h4 class=\"card-title show_txt\"><a href=\"detail?pid="+o.getProductId()+"\" title=\"View Product\">"+o.getProductName()+"</a></h4>\n"
-                    + "                                        <p class=\"card-text show_txt\">"+o.getDescription()+"</p>\n"
-                    + "                                        <div class=\"row\">\n"
-                    + "                                            <div class=\"col\">\n"
-                    + "                                                <p class=\"btn btn-danger btn-block\">"+o.getPrice()+" $</p>\n"
-                    + "                                            </div>\n"
-                    + "                                            <div class=\"col\">\n"
-                    + "                                                <a href=\"#\" class=\"btn btn-success btn-block\">Add to cart</a>\n"
-                    + "                                            </div>\n"
-                    + "                                        </div>\n"
-                    + "                                    </div>\n"
-                    + "                                </div>\n"
-                    + "                            </div>");
-        }
-    }
-
-    /**
-     * Handles the HTTP <code>POST</code> method.
-     *
-     * @param request servlet request
-     * @param response servlet response
-     * @throws ServletException if a servlet-specific error occurs
-     * @throws IOException if an I/O error occurs
-     */
-    @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        processRequest(request, response);
-    }
-
-    /**
-     * Returns a short description of the servlet.
-     *
-     * @return a String containing servlet description
-     */
-    @Override
-    public String getServletInfo() {
-        return "Short description";
-    }// </editor-fold>
-
-}
